+++ lib/project_types/extension/models/server_config/capabilities.rb
@@ -1,0 +1,11 @@
+module Extension
+  module Models
+    module ServerConfig
+      class Capabilities < Base
+        include SmartProperties
+
+        property :network_access, accepts: [true, false], default: false
+      end
+    end
+  end
+end